+++ prism-agent/service/server/src/main/scala/io/iohk/atala/agent/server/jobs/ConnectBackgroundJobs.scala
@@ -18,16 +18,21 @@ import io.iohk.atala.resolvers.UniversalDidResolver
 import io.iohk.atala.agent.walletapi.service.ManagedDIDService
 import io.iohk.atala.agent.walletapi.model.error.DIDSecretStorageError
 import io.iohk.atala.agent.walletapi.model.error.DIDSecretStorageError.KeyNotFoundError
+import io.iohk.atala.agent.server.config.AppConfig
 
 object ConnectBackgroundJobs {
 
   val didCommExchanges = {
     for {
       connectionService <- ZIO.service[ConnectionService]
+      config <- ZIO.service[AppConfig]
       records <- connectionService
-        .getConnectionRecords()
-        .mapError(err => Throwable(s"Error occured while getting connection records: $err"))
-      _ <- ZIO.foreach(records)(performExchange)
+        .getConnectionRecordsByStates(
+          ConnectionRecord.ProtocolState.ConnectionRequestPending,
+          ConnectionRecord.ProtocolState.ConnectionResponsePending
+        )
+        .mapError(err => Throwable(s"Error occurred while getting connection records: $err"))
+      _ <- ZIO.foreachPar(records)(performExchange).withParallelism(config.connect.connectBgJobProcessingParallelism)
     } yield ()
   }
 
@@ -54,9 +59,12 @@ object ConnectBackgroundJobs {
         val aux = for {
 
           didCommAgent <- buildDIDCommAgent(request.from)
-          _ <- MessagingService.send(request.makeMessage).provideSomeLayer(didCommAgent)
+          resp <- MessagingService.send(request.makeMessage).provideSomeLayer(didCommAgent)
           connectionService <- ZIO.service[ConnectionService]
-          _ <- connectionService.markConnectionRequestSent(id)
+          _ <- {
+            if (resp.status >= 200 && resp.status < 300) connectionService.markConnectionRequestSent(id)
+            else ZIO.logWarning(s"DIDComm sending error: [${resp.status}] - ${resp.bodyAsString}")
+          }
         } yield ()
 
         // aux // TODO decrete metaRetries if it has a error
@@ -78,9 +86,12 @@ object ConnectBackgroundJobs {
           ) if metaRetries > 0 =>
         val aux = for {
           didCommAgent <- buildDIDCommAgent(response.from)
-          _ <- MessagingService.send(response.makeMessage).provideSomeLayer(didCommAgent)
+          resp <- MessagingService.send(response.makeMessage).provideSomeLayer(didCommAgent)
           connectionService <- ZIO.service[ConnectionService]
-          _ <- connectionService.markConnectionResponseSent(id)
+          _ <- {
+            if (resp.status >= 200 && resp.status < 300) connectionService.markConnectionResponseSent(id)
+            else ZIO.logWarning(s"DIDComm sending error: [${resp.status}] - ${resp.bodyAsString}")
+          }
         } yield ()
 
         aux.tapError(ex =>